move endpoint to new revisions endpoint
