+++ src/server/src/controllers.rs
@@ -13,6 +13,7 @@ pub mod migrations;
 pub mod namespaces;
 pub mod not_found;
 pub mod repositories;
+pub mod revisions;
 pub mod schemas;
 pub mod stager;
 pub mod version;
